feat(ifame builder): introduces a new iframe builder for easier integration.
Either use the iframe element directly, with the client class to communicate with it. Or use the builder! The builder allows for a more programmatic approach to integrating Virtual Display. Project structure was totally changed and organized more logically. We have a created dedicated folders for iframe and message handling. For messaging new types have been introduced that hint at what direction the information flows. There are dedicated request and response objects and types now. It's possible to hook into the onResponse and onReady events sent out by the VirtualDisplayClient class. Through the builder with options too. Includes other minor tweaks.